Renamed 'RandomDice'
